backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.109.2
h11==0.16.0
idna==3.10
motor==3.3.2
passlib==1.7.4
prometheus-client==0.19.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.6.1
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
structlog==24.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.27.1
